Implement background and foreground
The background consists of "white" squares that have either not been
visited by the ant, or have been flipped back to the default color. The
foreground consists of "black" squares that have been visited by the ant
and flipped.

A list keeps a record of the flipped points that are supposed to be
black. Points are added and removed from this list.

The rendered grid is infinite.